requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
langchain_community 
langchain_pinecone
langchain-astradb
langchain-google-genai
flask
pypdf
python-dotenv
ipykernel 
pinecone-text 
pinecone-notebooks
python-dotenv
Cython
tiktoken
-e .
